Refactor source code and add .gitignore
