+++ .gitignore
@@ -1,0 +1,2 @@
+/src/__pycache__
+/test/__pycache__
